@@ -1,42 +1,40 @@
 {
-  "name": "angular-level1-tps",
-  "version": "0.1.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "build:prod": "ng build --prod",
-    "test": "ng test",
-    "lint": "ng lint",
-    "mock-apis": "json-server mocks/db.js --port 8585",
-    "doc:serve": "npx compodoc -p tsconfig.app.json -w -s",
-    "doc:build": "npx compodoc -p tsconfig.app.json"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "~8.0.0",
-    "@angular/common": "~8.0.0",
-    "@angular/compiler": "~8.0.0",
-    "@angular/core": "~8.0.0",
-    "@angular/forms": "~8.0.0",
-    "@angular/platform-browser": "~8.0.0",
-    "@angular/platform-browser-dynamic": "~8.0.0",
-    "@angular/router": "~8.0.0",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular/cli": "~8.0.2",
-    "@angular/compiler-cli": "~8.0.0",
-    "@angular/language-service": "~8.0.0",
-    "@compodoc/compodoc": "^1.1.11",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
-    "json-server": "^0.15.1",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.4.3"
-  }
+    "name": "angular-level1-tps",
+    "version": "0.1.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "build:prod": "ng build --prod",
+        "test": "ng test",
+        "lint": "ng lint",
+        "mock-apis": "json-server mocks/db.js --port 8585",
+        "doc:serve": "npx compodoc -p tsconfig.app.json -w -s",
+        "doc:build": "npx compodoc -p tsconfig.app.json"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "~11.0.0",
+        "@angular/common": "~11.0.0",
+        "@angular/compiler": "~11.0.0",
+        "@angular/core": "~11.0.0",
+        "@angular/forms": "~11.0.0",
+        "@angular/platform-browser": "~11.0.0",
+        "@angular/platform-browser-dynamic": "~11.0.0",
+        "@angular/router": "~11.0.0",
+        "rxjs": "~6.6.3",
+        "tslib": "^2.0.3",
+        "zone.js": "~0.11.3"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "~0.1100.0",
+        "@angular/cli": "~11.0.0",
+        "@angular/compiler-cli": "~11.0.0",
+        "@compodoc/compodoc": "^1.1.11",
+        "@types/node": "~14.14.7",
+        "json-server": "^0.16.2",
+        "ts-node": "~9.0.0",
+        "tslint": "~6.1.3",
+        "typescript": "~4.0.5"
+    }
 }
