@@ -4,14 +4,10 @@ import { DetailComponent } from './detail.component';
 
 const detailRoutes: Routes = [
   {
-    path: ':id',
+    path     : ':id',
     component: DetailComponent
   },
-  /*{
-    path: 'detail/:id',
-    component: DetailComponent
-  },*/
 ];
 
-export const detailRoute: ModuleWithProviders = RouterModule.forChild(detailRoutes);
+export const detailRoute: ModuleWithProviders<RouterModule> = RouterModule.forChild(detailRoutes);
 
