@@ -6,9 +6,9 @@ import { ComicModel } from '../shared/comics/comic.model';
 import { ComicsService } from '../shared/comics/comics.service';
 
 @Component({
-  selector: 'comics-detail',
+  selector   : 'comics-detail',
   templateUrl: './detail.component.html',
-  styleUrls: [ './detail.component.scss' ]
+  styleUrls  : ['./detail.component.scss']
 })
 export class DetailComponent implements OnInit {
   comic$: Observable<ComicModel>;
@@ -16,7 +16,8 @@ export class DetailComponent implements OnInit {
   constructor (
     private comicsService: ComicsService,
     private activatedRoute: ActivatedRoute
-  ) { }
+  ) {
+  }
 
   ngOnInit () {
     this.comic$ = this.comicsService.get(this.activatedRoute.snapshot.params.id);
